+++ QFuzzyMatcher/score_match.h
@@ -6,7 +6,7 @@
 struct MatchOptions {
   bool case_sensitive;
   bool smart_case;
-  size_t max_gap;
+  int max_gap;
 };
 
 /**
